PythonApi/requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.1.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Werkzeug==2.3.6
WTForms==3.0.1
